@@ -115,15 +115,16 @@ public class AboutActivity extends Activity {
             markdown = new Bypass(host, new Bypass.Options());
         }
 
+        @NonNull
         @Override
-        public Object instantiateItem(ViewGroup collection, int position) {
+        public Object instantiateItem(@NonNull ViewGroup collection, int position) {
             View layout = getPage(position, collection);
             collection.addView(layout);
             return layout;
         }
 
         @Override
-        public void destroyItem(ViewGroup collection, int position, Object view) {
+        public void destroyItem(@NonNull ViewGroup collection, int position, @NonNull Object view) {
             collection.removeView((View) view);
         }
 
@@ -133,7 +134,7 @@ public class AboutActivity extends Activity {
         }
 
         @Override
-        public boolean isViewFromObject(View view, Object object) {
+        public boolean isViewFromObject(@NonNull View view, @NonNull Object object) {
             return view == object;
         }
 
@@ -250,19 +251,16 @@ public class AboutActivity extends Activity {
         private @NonNull LibraryHolder createLibraryHolder(ViewGroup parent) {
             final LibraryHolder holder = new LibraryHolder(LayoutInflater.from(parent.getContext())
                     .inflate(R.layout.library, parent, false));
-            View.OnClickListener clickListener = new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    int position = holder.getAdapterPosition();
-                    if (position == RecyclerView.NO_POSITION) return;
-                    CustomTabActivityHelper.openCustomTab(
-                            host,
-                            new CustomTabsIntent.Builder()
-                                    .setToolbarColor(ContextCompat.getColor(host, R.color.primary))
-                                    .addDefaultShareMenuItem()
-                                    .build(), Uri.parse(libs[position - 1].link));
+            View.OnClickListener clickListener = v -> {
+                int position = holder.getAdapterPosition();
+                if (position == RecyclerView.NO_POSITION) return;
+                CustomTabActivityHelper.openCustomTab(
+                        host,
+                        new CustomTabsIntent.Builder()
+                                .setToolbarColor(ContextCompat.getColor(host, R.color.primary))
+                                .addDefaultShareMenuItem()
+                                .build(), Uri.parse(libs[position - 1].link));
 
-                }
             };
             holder.itemView.setOnClickListener(clickListener);
             holder.link.setOnClickListener(clickListener);
